100% INTs

+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: ycyr-roy <ycyr-roy@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/14 17:22:49 by ycyr-roy          #+#    #+#             */
-/*   Updated: 2023/03/30 21:23:14 by ycyr-roy         ###   ########.fr       */
+/*   Updated: 2023/04/04 13:57:44 by ycyr-roy         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -46,14 +46,14 @@ int			ft_putnbr_unsigned(unsigned int n, int *len);
 int			ft_atoi(const char *str);
 int			ft_printf(const char *str, ...);
 void		ft_strct_clean(t_data *data);
-int			int_len(int n);
+int			int_len(long long n);
 int			ft_atoi(const char *str);
 int			ft_flood(int size, char c, int *len);
 int			ft_strlen(const char *s);
 char		*ft_strdup(const char *s1);
 int			int_disp0(long long nb, t_data *data);
-int			int_disp1(int nb, int nb_len, t_data *data);
-int			int_disp2(int nb, int nb_len, t_data *data);
+int			int_disp1(long long nb, int nb_len, t_data *data);
+int			int_disp2(long long nb, int nb_len, t_data *data);
 int			int_disp3(int nb_len, t_data *data);
 int			c_dispatcher(int nb, t_data *data);
 int			d_dispatcher(int nb, t_data *data);
